% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{df_rename}
\alias{df_rename}
\title{Illustration of crayon colors}
\usage{
df_rename(data_set, names, title)
}
\arguments{
\item{data_set}{The data frame to be edited}

\item{names}{the rownames to be displayed}

\item{title}{the names of the row name.}
}
\value{
None
}
\description{
Creates a plot of the crayon colors in \code{\link{brocolors}}
}
\examples{
df_rename(
     data.frame(letters[1:5], 1:5),
     LETTERS[1:5],
     "CAPS"
)

}

