% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{nice_table}
\alias{nice_table}
\title{Pretty tables for LM}
\usage{
nice_table(fit, level = 0.95, roundto = 2)
}
\arguments{
\item{fit}{The lm object}

\item{level}{confidence level}

\item{roundto}{how much do you want the results rounded to}
}
\value{
None
}
\description{
Creates a knitr approved table with regression coefficients and confidence intervals
}
\examples{
fit <- lm(nrorm(100) ~ rnorm(100,2,3))
nice_table(fit, 0.90, roundto = 3)

}

