% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ollama_chat}
\alias{ollama_chat}
\title{Chat}
\usage{
ollama_chat(ollama, model, query, previous_messages = NULL, return_all = FALSE)
}
\arguments{
\item{ollama}{Imported module from R}

\item{model}{ollama model to be used}

\item{return_all}{return the whole output (default is FALSE)}
}
\description{
Chat
}
