% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ollama_generate}
\alias{ollama_generate}
\title{Generate}
\usage{
ollama_generate(ollama, model, query, return_all = FALSE)
}
\arguments{
\item{ollama}{Imported module from R}

\item{model}{ollama model to be used}

\item{return_all}{return the whole output (default is FALSE)}
}
\description{
Generate
}
